+++ orthoseg/__init__.py
@@ -7,7 +7,6 @@ from pathlib import Path
 # ruff: noqa: F401
 from orthoseg.load_images import load_images
 from orthoseg.train import train
-from orthoseg.train import _search_label_files
 from orthoseg.predict import predict
 from orthoseg.postprocess import postprocess