+++ orthoseg/lib/prepare_traindatasets.py
@@ -39,6 +39,8 @@ class LabelInfo:
         locations_path: Path,
         polygons_path: Path,
         image_layer: str,
+        pixel_x_size: Optional[float] = None,
+        pixel_y_size: Optional[float] = None,
         locations_gdf: Optional[gpd.GeoDataFrame] = None,
         polygons_gdf: Optional[gpd.GeoDataFrame] = None,
     ):
@@ -49,6 +51,8 @@ class LabelInfo:
             locations_path (Path): file path to the locations file.
             polygons_path (Path): file path to the polygons file.
             image_layer (str): image layer for the label.
+            pixel_x_size (Optional[float], optional): _description_. Defaults to None.
+            pixel_y_size (Optional[float], optional): _description_. Defaults to None.
             locations_gdf (Optional[gpd.GeoDataFrame], optional): _description_.
                 Defaults to None.
             polygons_gdf (Optional[gpd.GeoDataFrame], optional): _description_.
@@ -57,6 +61,8 @@ class LabelInfo:
         self.locations_path = locations_path
         self.polygons_path = polygons_path
         self.image_layer = image_layer
+        self.pixel_x_size = pixel_x_size
+        self.pixel_y_size = pixel_y_size
         self.locations_gdf = locations_gdf
         self.polygons_gdf = polygons_gdf
 
@@ -383,8 +389,8 @@ def prepare_labeldata(
     label_infos: List[LabelInfo],
     classes: dict,
     labelname_column: str,
-    image_pixel_x_size: float,
-    image_pixel_y_size: float,
+    image_pixel_x_size: Optional[float],
+    image_pixel_y_size: Optional[float],
     image_pixel_width: int,
     image_pixel_height: int,
 ) -> List[Tuple[gpd.GeoDataFrame, gpd.GeoDataFrame]]:
@@ -399,10 +405,10 @@ def prepare_labeldata(
             label class names + weights.
         labelname_column (str): the column name in the label polygon files where the
             label classname can be found. Defaults to "classname".
-        image_pixel_x_size (float): pixel size in x direction. Defaults to 0.25.
-        image_pixel_y_size (float): pixel size in y direction. Defaults to 0.25.
-        image_pixel_width (int): width of the image in pixels. Defaults to 512.
-        image_pixel_height (int): height of the image in pixels. Defaults to 512.
+        image_pixel_x_size (float): pixel size in x direction.
+        image_pixel_y_size (float): pixel size in y direction.
+        image_pixel_width (int): width of the image in pixels.
+        image_pixel_height (int): height of the image in pixels.
 
     Raises:
         ValidationError: if the data is somehow not valid. All issues are listed in the
@@ -463,12 +469,16 @@ def prepare_labeldata(
 
         # Check + correct the geom of the location to make sure it matches the size of
         # the training images/masks to be generated...
-        image_crs_width = math.fabs(
-            image_pixel_width * image_pixel_x_size
-        )  # tile width in units of crs => 500 m
-        image_crs_height = math.fabs(
-            image_pixel_height * image_pixel_y_size
-        )  # tile height in units of crs => 500 m
+        # Prepare pixel size: if defined on label info level, use that.
+        pixel_x_size = label_info.pixel_x_size
+        pixel_x_size = pixel_x_size if pixel_x_size is not None else image_pixel_x_size
+        pixel_y_size = label_info.pixel_y_size
+        pixel_y_size = pixel_y_size if pixel_y_size is not None else image_pixel_y_size
+
+        # Tile width/height in units of crs
+        image_crs_width = math.fabs(image_pixel_width * pixel_x_size)
+        image_crs_height = math.fabs(image_pixel_height * pixel_y_size)
+
         locations_none = []
         for location in labellocations_gdf.itertuples():
             if location.geometry is None or location.geometry.is_empty: